requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
azure-ai-translation-text==1.0.1
azure-core==1.32.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
cryptography==44.0.0
deepl==1.20.0
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
flatbuffers==24.12.23
fsspec==2024.12.0
gradio==5.10.0
gradio_client==1.5.3
gradio_pdf==0.0.21
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.5
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
numpy==2.2.1
ollama==0.4.5
onnx==1.17.0
onnxruntime==1.20.1
openai==1.59.3
opencv-python-headless==4.10.0.84
orjson==3.10.13
packaging==24.2
pandas==2.2.3
pdf2zh==1.8.8
pdfminer.six==20240706
pillow==11.1.0
protobuf==5.29.2
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyMuPDF==1.25.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-telegram-bot==21.10
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.8.6
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
tencentcloud-sdk-python==3.0.1297
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.1
